+++ casia/src/components/atoms/Button/Button.js
@@ -1,4 +1,4 @@
-import styled from 'styled-components';
+import styled, { css } from 'styled-components';
 
 const Button = styled.button`
   padding: 10px 20px;
@@ -10,6 +10,15 @@ const Button = styled.button`
   border-radius: 20px;
   text-transform: uppercase;
   font-weight: 700;
+
+  ${({ secondary }) =>
+    secondary &&
+    css`
+      background-color: grey;
+      width: 105px;
+      height: 30px;
+      font-size: 10px;
+    `}
 `;
 
 export default Button;