+++ casia/src/components/molecules/Note/note.js
@@ -1,11 +1,17 @@
 import React from 'react';
 import styled from 'styled-components';
+import { connect } from 'react-redux';
+import { deleteNote, noteContentHandler } from '../../../actions/index';
+
+import Button from '../../atoms/Button/Button';
 
 const Wrapper = styled.div`
   border-radius: 20px;
   border: 2px solid #ccc;
   margin: 5px;
   overflow: hidden;
+  display: flex;
+  flex-direction: column;
 `;
 
 const Header = styled.input`
@@ -25,13 +31,33 @@ const CardContent = styled.textarea`
   border: none;
 `;
 
+const ButtonWrapper = styled.div`
+  width: 100%;
+  text-align: center;
+`;
+
 const Note = props => {
   return (
     <Wrapper>
-      <Header value={props.title}></Header>
-      <CardContent value={props.context} />
+      <Header
+        value={props.title}
+        onChange={e =>
+          props.dispatch(noteContentHandler(e.target.value, props.id, 'TITLE'))
+        }
+      ></Header>
+      <CardContent
+        onChange={e =>
+          props.dispatch(noteContentHandler(e.target.value, props.id))
+        }
+        value={props.context}
+      />
+      <ButtonWrapper>
+        <Button secondary onClick={e => props.dispatch(deleteNote(props.id))}>
+          DELETE
+        </Button>
+      </ButtonWrapper>
     </Wrapper>
   );
 };
 
-export default Note;
+export default connect(null, null)(Note);