+++ casia/src/components/molecules/SidebarItem/SidebarItem.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import ButtonItem from '../../atoms/ButtonIcon/ButtonIcon';
+
+const SidebarItem = props => <ButtonItem icon={props.icon}></ButtonItem>;
+
+export default SidebarItem;