+++ casia/src/routes/routes.js
@@ -1,0 +1,7 @@
+export default {
+  homePage: '/',
+  notesPage: '/Notes',
+  toDoPage: './Todo',
+  imagePage: './Images',
+  newItemPage: './CreateItem'
+};