+++ casia/src/views/CreateItemView.js
@@ -5,8 +5,6 @@ import MainTemplate from '../components/templates/MainTemplate/MainTemplate';
 import Form from '../components/organisms/Form/Form';
 
 const CreateItemView = props => {
-  console.log(props);
-
   return (
     <MainTemplate>
       <h1>You Choose: {props.creationItem}</h1>