+++ casia/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change log
 
+## V.0.1.2 Saturday, 14 March 2020
+
+**Added**
+
+- Deleting notes and modyfing them.
+- Routes.js with all paths
+
 ## V.0.1.1 Friday, 13 March 2020
 
 **Added**
@@ -7,7 +14,7 @@
 - Redux, React-Redux (reducer, actions, store)
 - CreateItemView,
 - Form to add notes,
-- Input/Textarea, Form, components,
+- Input/Textarea, Form, Cards components,
 - Constants with colors
 
 Initial relase.